refactored views, some tests still failing
