+++ tardis/urls.py
@@ -38,15 +38,16 @@ rapidconnect_urls = patterns(
 core_urls = patterns(
     'tardis.tardis_portal.views',
     (r'^$', 'index'),
-    url(r'^site-settings.xml/$', 'site_settings', name='tardis-site-settings'),
+    url(r'^site-settings.xml/$',
+        'machine.site_settings', name='tardis-site-settings'),
     url(r'^mydata/$', 'my_data', name='mydata'),
     url(r'^public_data/', 'public_data', name='public_data'),
     (r'^about/$', 'about'),
     (r'^stats/$', 'stats'),
     (r'^help/$', 'user_guide'),
     url(r'^sftp_access/cyberduck/connection.png$',
-        'cybderduck_connection_window', name='cyberduck_connection_window'),
-    url(r'^sftp_access/$', 'sftp_access', name='sftp_access'),
+        'images.cybderduck_connection_window', name='cyberduck_connection_window'),
+    url(r'^sftp_access/$', 'pages.sftp_access', name='sftp_access'),
     (r'^robots\.txt$', lambda r: HttpResponse(
         "User-agent: *\nDisallow: /download/\nDisallow: /stats/",
         content_type="text/plain"))
@@ -169,7 +170,8 @@ ajax_urls = patterns(
     (r'^experiment_metadata/(?P<experiment_id>\d+)/$',
         'retrieve_experiment_metadata'),
     (r'^datafile_list/(?P<dataset_id>\d+)/$', 'retrieve_datafile_list'),
-    url(r'^cache_dataset/(?P<dataset_id>\d+)/$', 'cache_dataset',
+    url(r'^cache_dataset/(?P<dataset_id>\d+)/$',
+        'ajax_actions.cache_dataset',
         name='cache_dataset'),
     (r'^user_list/$', 'retrieve_user_list'),
     (r'^group_list/$', 'retrieve_group_list'),
